requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
google-generativeai>=0.3.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pillow>=10.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
